pyproject.toml
[project]
name = "english-buddy"
version = "0.1.0"
description = "Toddler English learning assistant for Chinese-speaking parents"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.0.0",
    "pydantic>=2.0",
    "rich>=13.0",
    "python-dotenv>=1.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
]

[project.scripts]
english-buddy = "english_buddy.__main__:main"
english-buddy-web = "english_buddy.api:main"
